web/maintainer: Move /finance into /maintainer
