+++ clients/apps/web/src/pages/finance/[organization]/rewarded.tsx
@@ -1,13 +1,7 @@
 import Gatekeeper from '@/components/Dashboard/Gatekeeper/Gatekeeper'
-import Transactions from '@/components/Finance/Finance'
-import DashboardLayout from '@/components/Layout/DashboardLayout'
+import LoadingScreen from '@/components/Dashboard/LoadingScreen'
 import type { NextLayoutComponentType } from 'next'
-import Head from 'next/head'
 import { useRouter } from 'next/router'
-import {
-  useListPledgesForOrganization,
-  useOrganizationAccounts,
-} from 'polarkit/hooks'
 import { ReactElement, useEffect } from 'react'
 import { useCurrentOrgAndRepoFromURL } from '../../../hooks'
 
@@ -16,39 +10,26 @@ const Page: NextLayoutComponentType = () => {
   const { org, isLoaded } = useCurrentOrgAndRepoFromURL()
 
   useEffect(() => {
-    if (isLoaded && !org) {
-      router.push('/issues')
+    if (!isLoaded) return
+
+    if (org) {
+      router.push(`/maintainer/${org.name}/finance/rewarded`)
       return
     }
+    router.push(`/dependencies/personal`)
   }, [isLoaded, org, router])
-
-  const pledges = useListPledgesForOrganization(org?.platform, org?.name)
-
-  const accounts = useOrganizationAccounts(org?.name)
 
   return (
     <>
-      <Head>
-        <title>Polar{org ? ` ${org.name}` : ''}</title>
-      </Head>
-      {org && pledges.data && accounts.data && (
-        <Transactions
-          pledges={pledges.data}
-          org={org}
-          tab="rewarded"
-          accounts={accounts.data}
-        />
-      )}
+      <LoadingScreen>
+        <>Redirecting...</>
+      </LoadingScreen>
     </>
   )
 }
 
 Page.getLayout = (page: ReactElement) => {
-  return (
-    <Gatekeeper>
-      <DashboardLayout showSidebar={true}>{page}</DashboardLayout>
-    </Gatekeeper>
-  )
+  return <Gatekeeper>{page}</Gatekeeper>
 }
 
 export default Page
